Init Repo
Basically created a basic project. i havent run the docker-compose file to create an environment but i just put it in. Also there arent any comments yet I'll still do all of that this is just there so that people who are working on different things kinda get an idea of how things run.